Move tests to UserSession
